feat: SQL bases streamStore: factor implementation out of tests
